+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## Unreleased
 
 - Minor: Add basic support for Fluxer webhooks. (#896)
+- Minor: Add Demonic Brutus as a drop source to Beef for luck calculation. (#897)
+- Minor: Add `%PET%` template variable to inject the obtained pet name if known. (#899)
 - Dev: Send GIM shared bank contents to the custom metadata webhook upon opening the storage interface. (#895)
 
 ## 1.12.1